backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
PyMuPDF==1.24.10
openai==1.51.0
tiktoken==0.7.0
python-multipart==0.0.12
python-dotenv==1.0.1
